+++ src/main/java/roomescape/service/dto/request/ReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto;
+package roomescape.service.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import java.time.LocalTime;
@@ -6,7 +6,6 @@ import roomescape.controller.exception.TimeValid;
 import roomescape.domain.ReservationTime;
 
 public record ReservationTimeRequest(
-
         @NotBlank(message = "예약 시작 시간을 입력해주세요.")
         @TimeValid
         String startAt
