+++ src/main/java/roomescape/service/dto/response/AvailableReservationTimeResponse.java
@@ -1,12 +1,12 @@
-package roomescape.service.dto;
+package roomescape.service.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
+import roomescape.domain.ReservationTime;
 
 public record AvailableReservationTimeResponse(
-        Long timeId,
+        Long id,
         @JsonFormat(pattern = "HH:mm")
         LocalTime startAt,
         boolean alreadyBooked
-) {
-}
+) {}
