+++ src/test/java/roomescape/controller/ThemeControllerTest.java
@@ -19,8 +19,8 @@ import org.springframework.test.context.jdbc.Sql;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
-import roomescape.service.dto.ThemeRequest;
-import roomescape.service.dto.ThemeResponse;
+import roomescape.service.dto.request.ThemeRequest;
+import roomescape.service.dto.response.ThemeResponse;
 import roomescape.domain.ReservationRepository;
 import roomescape.domain.ReservationTimeRepository;
 import roomescape.domain.ThemeRepository;
